feat: Addressed #7 and refactored transform.js in consequence
