+++ gulpfile.js/index.js
@@ -4,6 +4,7 @@ const transform = require('gulp-transform');
 const rename = require('gulp-rename');
 
 const {
+  classifyTools,
   validateMetadata,
   getRepositoryMetadata,
   readLocalSourceData,
@@ -12,6 +13,8 @@ const {
   normaliseSources,
 } = require('../lib');
 
+// This is complete scan of the source data. All sources will be retrieved and processed
+// though the transformation code, rebuilding tools.yaml
 const full = () => src('gulpfile.js/metadata.json')
   .pipe(transform('utf8', validateMetadata))
   .pipe(transform('utf8', readSourceData))
@@ -20,9 +23,12 @@ const full = () => src('gulpfile.js/metadata.json')
   .pipe(transform('utf8', mergeSources))
   .pipe(transform('utf8', normaliseSources))
   .pipe(transform('utf8', getRepositoryMetadata))
+  .pipe(transform('utf8', classifyTools))
   .pipe(rename('tools.yaml'))
   .pipe(dest('docs/'));
 
+// This is a scan of the metadata associated with the repositories already
+// held in the repository. No new source data is retrieved from sources
 const metadata = () => src('gulpfile.js/metadata.json')
   .pipe(transform('utf8', validateMetadata))
   .pipe(transform('utf8', readLocalSourceData))
