+++ lib/index.js
@@ -1,4 +1,5 @@
 const {
+  classifyTools,
   validateMetadata,
   getRepositoryMetadata,
   readLocalSourceData,
@@ -8,6 +9,7 @@ const {
 } = require('./transform');
 
 module.exports = {
+  classifyTools,
   validateMetadata,
   getRepositoryMetadata,
   readLocalSourceData,
