+++ lib/processors/openapi-implementations-md-processor.js
@@ -2,7 +2,13 @@ const http = require('axios');
 
 const { logger } = require('../util');
 
-module.exports = async (source, url) => {
+module.exports = async (args) => {
+  const { title, url } = args;
+
+  if (!title || !url) {
+    throw new Error(`Mandatory parameters missing when invoking ${__filename}`);
+  }
+
   const response = await http.get(url, { timeout: 5000 });
   let toolType;
 
@@ -21,7 +27,7 @@ module.exports = async (source, url) => {
         .map((attribute) => attribute.trim());
 
       return output.concat({
-        source,
+        source: title,
         name: properties[0],
         homepage: properties[1].match(/http.*[a-z0-9A-Z/]+/).pop(),
         languages: properties[2]
@@ -32,7 +38,7 @@ module.exports = async (source, url) => {
       });
     }, []);
 
-  logger('Number of sources found in IMPLEMENTATIONS.md', tools.length);
+  logger(`Number of sources found in ${title}`, tools.length);
 
   return tools;
 };
