+++ lib/processors/openapi-tools-processor.js
@@ -3,11 +3,17 @@ const YAML = require('js-yaml');
 
 const { logger } = require('../util');
 
-module.exports = async (source, url) => {
+module.exports = async (args) => {
+  const { title, url } = args;
+
+  if (!title || !url) {
+    throw new Error(`Mandatory parameters missing when invoking ${__filename}`);
+  }
+
   const response = await http.get(url, { timeout: 5000 });
   const data = YAML.load(response.data);
 
-  logger('Number of sources found in openapi.tools', data.length);
+  logger(`Number of sources found in ${title}`, data.length);
 
-  return data.map((tool) => ({ source, ...tool }));
+  return data.map((tool) => ({ source: title, ...tool }));
 };
