+++ lib/processors/tagged-repository-processor.js
@@ -2,9 +2,9 @@ const Graphql = require('graphql-request').GraphQLClient;
 
 const { logger } = require('../util');
 
-const wrapper = async (source, client, results, recursor) => {
+const wrapper = async (client, topic, results, recursor) => {
   const query = `{
-    search(type: REPOSITORY, query: "topic:openapi3", first: 100, ${recursor ? `after: "${recursor}"` : ''}) {
+    search(type: REPOSITORY, query: "topic:${topic}", first: 100, ${recursor ? `after: "${recursor}"` : ''}) {
       edges{
         node {
           ... on Repository {
@@ -35,31 +35,36 @@ const wrapper = async (source, client, results, recursor) => {
     });
 
   if (lastCursor) {
-    await wrapper(source, client, results, lastCursor);
+    await wrapper(client, topic, results, lastCursor);
   }
 
   return results;
 };
 
-module.exports = async (source) => {
+module.exports = async (args) => {
   const results = [];
+  const { title, topic, version } = args;
+
+  if (!title || !topic || !version) {
+    throw new Error(`Mandatory parameters missing when invoking ${__filename}`);
+  }
 
   await wrapper(
-    source,
     new Graphql('https://api.github.com/graphql', {
       headers: {
-        Authorization: `Basic ${Buffer.from(`${process.env.GITHUB_ACTOR}:${process.env.GITHUB_TOKEN}`).toString('base64')}`,
+        Authorization: `Basic ${Buffer.from(`${process.env.GH_API_USERNAME}:${process.env.GH_API_TOKEN}`).toString('base64')}`,
       },
     }),
+    topic,
     results,
   );
 
   // Despite the pagination the GitHub query appears to return duplicates
   // Use a Set to dedup before returning with additional metadata
   const output = [...new Set(results)]
-    .map((url) => ({ source, github: url, v3: true }));
+    .map((url) => ({ source: title, github: url, [version]: true }));
 
-  logger('Number of sources found tagged with openapi3', output.length);
+  logger(`Number of sources found tagged with ${topic}`, output.length);
 
   return output;
 };
