+++ lib/util.js
@@ -3,6 +3,7 @@ const dl = require('talisman/metrics/damerau-levenshtein');
 const log = require('fancy-log');
 
 module.exports = {
+  getUri: (source) => (source.repository || source.github || source.link || source.homepage),
   logger: (...args) => {
     if (process.env.NODE_ENV !== 'test') {
       log(...args);
