+++ package.json
@@ -17,6 +17,7 @@
     "@11ty/eleventy": "^1.0",
     "ajv": "^8.10.0",
     "axios": "^0.24.0",
+    "bayes": "^1.0.0",
     "fancy-log": "^1.3.3",
     "graphql": "14",
     "graphql-request": "^3.7.0",
